+++ src/app/modules/showcase-ui/examples/button-example/button-example.component.ts
@@ -1,11 +1,11 @@
-import { Component } from '@angular/core';
-import { Lab900ActionButtonComponent, Lab900ButtonComponent, ToggleActionButton } from '@lab900/ui';
-import { of } from 'rxjs';
+import { ChangeDetectionStrategy, Component } from '@angular/core';
+import { Lab900ButtonComponent } from '@lab900/ui';
 
 @Component({
   selector: 'lab900-button-example',
   styles: ['p {margin: 10px 0}'],
-  imports: [Lab900ButtonComponent, Lab900ActionButtonComponent],
+  imports: [Lab900ButtonComponent],
+  changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
     <p>Raised button</p>
     <lab900-button
@@ -15,10 +15,10 @@ import { of } from 'rxjs';
       label="hello world"
       suffixIcon="remove_red_eye"
       buttonId="Raised-button"
-      throttleTimeInMs="1000" />
+      [throttleTimeInMs]="1000" />
     <p>Raised button disabled</p>
     <lab900-button
-      disabled="true"
+      [disabled]="true"
       (btnClick)="log('Raised button disabled')"
       color="primary"
       type="raised"
@@ -36,7 +36,7 @@ import { of } from 'rxjs';
     <p>Stroked button disabled</p>
     <lab900-button
       (btnClick)="log('Stroked button disabled')"
-      disabled="true"
+      [disabled]="true"
       color="primary"
       type="stroked"
       label="hello world"
@@ -50,42 +50,9 @@ import { of } from 'rxjs';
     <lab900-button type="fab" label="delete" (btnClick)="log(' Fab Icon button')" />
     <p>Fab-mini icon button</p>
     <lab900-button type="mini-fab" label="delete" (btnClick)="log('Fab Mini Icon button')" />
-    <p>Toggle action button</p>
-    <lab900-action-button [data]="[]" [action]="toggleActionButton" />
   `,
 })
 export class ButtonExampleComponent {
-  public toggleActionButton: ToggleActionButton = {
-    label: 'EntityDetailSectionToggleToggleButton',
-    type: 'toggle',
-    subActions: [
-      {
-        label: 'Button 1',
-        selected: true,
-        action: () => {
-          console.log('Button 1 is selected');
-        },
-        buttonId: 'btn-1',
-      },
-      {
-        label: 'Button 2',
-        selected: false,
-        action: () => {
-          console.log('Button 2 is selected');
-        },
-        buttonId: 'btn-2',
-      },
-      {
-        label: 'Button 3',
-        action: () => {
-          console.log('Button 2 is selected');
-        },
-        hide: () => of(true),
-        buttonId: 'btn-3',
-      },
-    ],
-  };
-
   public log(message: string): void {
     console.log(message);
   }