+++ src/app/modules/showcase-ui/examples/page-header-example/page-header-example.component.ts
@@ -1,7 +1,6 @@
 import { Component } from '@angular/core';
 import { ActionButton, BreadCrumb, Lab900PageHeaderComponent, PageHeaderNavItem } from '@lab900/ui';
 import { MatTabsModule } from '@angular/material/tabs';
-import { of } from 'rxjs';
 
 @Component({
   selector: 'lab900-page-header-example',
@@ -56,7 +55,7 @@ export class PageHeaderExampleComponent {
     {
       type: 'toggle',
       label: '',
-      disabled: () => of(true),
+      disabled: true,
       subActions: [
         {
           label: 'Action 1',